update CI tests and terraform to support agented VMs (#42)
* Add support to import Agented VMs

- In Agented mode, VirtualMachine CR will be converted to ExternalNode
  CR and antrea-controller will create ExternalEntity CR.
- In Agentless mode, VirtualMachine CR will be converted to
  ExternalEntity CR.
- Added agented flag in CloudEntitySelector CRD.
- Added agented flag in VirtualMachine CRD.
- Update antrea.yml to enable ExternalNodeController
  and set antrea service type to LoadBalancer.
- Update converter module to create ExternalNode CR
  for agented VMs and create create ExternalEntity CR
  for agentless VMs.
- Implement cache to get vmSelector from VMid/VPCid
- Display Agented column for VM CRD

To support transition from agented to agentless, VirtualMachine CR will
be deleted and created, so that any dependent CRs will be cleaned
and recreated again.

Add unit tests for ExternalNode
- Converter source package coverage at 87.6%


* Add antrea-agent wrapper install script for cloud VMs

- Downloads the install script from antrea.
- Downloads the required files from a specific antrea release version.
- Generates the ExternalNode name and feeds to antrea install script.
- Supports both AWS and Azure cloud.
- Supports only Ubuntu and Windows OS.


* Add documentation for Agented VM


* terraform support aws agented vm


* open security for agented vm


* cluster config script and update terraform var


* update ci test with agent config and focus


* cleanup and update jenkins scripts


* lint files


* resolve comments


* add azure support for agented ci test


* add vm agent log collection


* update docs on ci tests, terraform and helper scripts


* new jenkins jobs for agented ci test


* reformat docs, comments and var names


* fix terraform var defaults


* change jenkins job tag and reformat


* Update jenkins jobs


* update jenkins jobs


* fix ci and resolve comments


* use tmp directory for generate kubeconfig in ci


* fix lint issue


* fix aks test script


* update antrea to 1.9 and remove custom images


* resolve comments


* update jenkins job trigger to avoid mistrigger


* Publish install scripts for vm agent as assets (#54)

* Publish install scripts for vm agent as assets

Rename and upload install-vm-agent-wrapper scripts as
assets.


* Update docs to download scripts from assets


* Handle script file name changes


* fix jenkins jobs typo


* update log msg


* delete nephe when ci finishes


* fix merge error


* Cleanup antrea resources when cluster is destroyed

+++ test/integration/cloudresource_e2e_test.go
@@ -45,10 +45,11 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 		apachePort = "8080"
 	)
 	var (
-		namespace      *v1.Namespace
-		otherNamespace *v1.Namespace
-		anpParams      k8stemplates.ANPParameters
-		anpSetupParams k8stemplates.ANPParameters
+		namespace            *v1.Namespace
+		otherNamespace       *v1.Namespace
+		anpParams            k8stemplates.ANPParameters
+		anpSetupParams       k8stemplates.ANPParameters
+		defaultANPParameters k8stemplates.DefaultANPParameters
 
 		importAfterANP bool
 		abbreviated    bool
@@ -70,7 +71,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			if len(supportBundleDir) > 0 {
 				logf.Log.Info("Collect support bundles for test failure.")
 				fileName := utils.GenerateNameFromText(result.FullTestText, testFocus)
-				utils.CollectSupportBundle(kubeCtl, path.Join(supportBundleDir, fileName))
+				utils.CollectSupportBundle(kubeCtl, path.Join(supportBundleDir, fileName), cloudVPC, withAgent)
 				bundleCollected = true
 			}
 			if preserveSetupOnFail {
@@ -83,17 +84,19 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			if !bundleCollected && err != nil && len(supportBundleDir) > 0 {
 				logf.Log.Info("Collect support bundles for clean-up failure.")
 				fileName := utils.GenerateNameFromText(result.FullTestText, testFocus)
-				utils.CollectSupportBundle(kubeCtl, path.Join(supportBundleDir, fileName))
+				utils.CollectSupportBundle(kubeCtl, path.Join(supportBundleDir, fileName), cloudVPC, withAgent)
 			}
 
-			nss := []*v1.Namespace{namespace, otherNamespace}
-			for _, ns := range nss {
-				if ns == nil {
-					continue
+			if !withAgent {
+				nss := []*v1.Namespace{namespace, otherNamespace}
+				for _, ns := range nss {
+					if ns == nil {
+						continue
+					}
+					logf.Log.V(1).Info("Delete namespace", "ns", ns.Name)
+					err := k8sClient.Delete(context.TODO(), ns)
+					Expect(err).ToNot(HaveOccurred())
 				}
-				logf.Log.V(1).Info("Delete namespace", "ns", ns.Name)
-				err := k8sClient.Delete(context.TODO(), ns)
-				Expect(err).ToNot(HaveOccurred())
 			}
 			preserveSetup = false
 		}()
@@ -103,10 +106,9 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 		Expect(err).ToNot(HaveOccurred())
 		err = utils.ConfigureK8s(kubeCtl, anpParams, k8stemplates.CloudAntreaNetworkPolicy, true)
 		Expect(err).ToNot(HaveOccurred())
-		err = utils.CheckCloudResourceNetworkPolicies(k8sClient, reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), namespace.Name,
-			cloudVPC.GetVMs(), nil)
+		err = utils.CheckCloudResourceNetworkPolicies(kubeCtl, k8sClient, reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), namespace.Name,
+			cloudVPC.GetVMs(), nil, withAgent)
 		Expect(err).ToNot(HaveOccurred())
-
 	})
 
 	configANPApplyTo = func(kind, instanceName, vpc, tagKey, tagVal string) *k8stemplates.EntitySelectorParameters {
@@ -157,6 +159,10 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 		if !diffNS {
 			otherNamespace = nil
 		}
+		if withAgent {
+			namespace = staticVMNS
+			otherNamespace = nil
+		}
 
 		nss := []*v1.Namespace{namespace, otherNamespace}
 		for _, ns := range nss {
@@ -178,9 +184,20 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 
 			if !importAfterANP {
 				entityParams := cloudVPC.GetEntitySelectorParameters("test-entity-selector"+ns.Name, ns.Name, kind)
+				entityParams.Agented = withAgent
 				err = utils.ConfigureEntitySelectorAndWait(kubeCtl, k8sClient, entityParams, kind, num, ns.Name, false)
 				Expect(err).ToNot(HaveOccurred())
 			}
+		}
+
+		if withAgent {
+			// To Keep tests happy, add default ingress/egress deny rule
+			// TODO: Update the tests later to remove this rule
+			defaultANPParameters = k8stemplates.DefaultANPParameters{
+				Namespace: staticVMNS.Name,
+			}
+			err := utils.ConfigureK8s(kubeCtl, defaultANPParameters, k8stemplates.DefaultANPSetup, false)
+			Expect(err).ToNot(HaveOccurred(), "failed to add default ANP rule")
 		}
 
 		anpParams = k8stemplates.ANPParameters{
@@ -212,11 +229,14 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 				}
 				_ = cloudVPC.GetCloudAccountParameters("test-cloud-account"+ns.Name, ns.Name, cloudCluster)
 				entityParams := cloudVPC.GetEntitySelectorParameters("test-entity-selector"+ns.Name, ns.Name, kind)
+				entityParams.Agented = withAgent
 				err = utils.ConfigureEntitySelectorAndWait(kubeCtl, k8sClient, entityParams, kind, num, ns.Name, false)
 				Expect(err).ToNot(HaveOccurred())
 			}
 		}
-		err = utils.CheckCloudResourceNetworkPolicies(k8sClient, vmKind, namespace.Name, cloudVPC.GetVMs(), []string{anpSetupParams.Name})
+		err = utils.CheckCloudResourceNetworkPolicies(
+			kubeCtl, k8sClient, vmKind, namespace.Name, cloudVPC.GetVMs(), []string{anpSetupParams.Name}, withAgent,
+		)
 		Expect(err).ToNot(HaveOccurred())
 		oks := make([]bool, len(cloudVPC.GetVMs()))
 		for i := range oks {
@@ -238,7 +258,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			if ok {
 				np = append(np, anpParams.Name)
 			}
-			err = utils.CheckCloudResourceNetworkPolicies(k8sClient, kind, namespace.Name, []string{ids[i]}, np)
+			err = utils.CheckCloudResourceNetworkPolicies(kubeCtl, k8sClient, kind, namespace.Name, []string{ids[i]}, np, withAgent)
 			Expect(err).ToNot(HaveOccurred())
 		}
 		appliedIPs := make([]string, 0)
@@ -275,7 +295,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			np = append(np, anpSetupParams.Name)
 		}
 		np = append(np, anpParams.Name)
-		err = utils.CheckCloudResourceNetworkPolicies(k8sClient, kind, namespace.Name, []string{id}, np)
+		err = utils.CheckCloudResourceNetworkPolicies(kubeCtl, k8sClient, kind, namespace.Name, []string{id}, np, withAgent)
 		Expect(err).ToNot(HaveOccurred())
 
 		err = utils.ExecuteCurlCmds(cloudVPC, nil, []string{srcVM}, "", dstIPs, apachePort, oks, 30)
@@ -294,7 +314,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			np = append(np, anpSetupParams.Name)
 		}
 		np = append(np, anpParams.Name)
-		err = utils.CheckCloudResourceNetworkPolicies(k8sClient, kind, namespace.Name, []string{id}, np)
+		err = utils.CheckCloudResourceNetworkPolicies(kubeCtl, k8sClient, kind, namespace.Name, []string{id}, np, withAgent)
 		Expect(err).ToNot(HaveOccurred())
 
 		err = utils.ExecuteCurlCmds(cloudVPC, nil, srcVMs, "", []string{ip}, apachePort, oks, 30)
@@ -542,7 +562,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 		func(kind string, diffNS bool) {
 			testAppliedTo(kind, diffNS)
 		},
-		table.Entry(fmt.Sprintf("%s: VM In Same Namespace", focusAzure),
+		table.Entry(fmt.Sprintf("%s %s: VM In Same Namespace", focusAzure, focusAgent),
 			reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), false),
 		table.Entry("VM In Different Namespaces",
 			reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), true),
@@ -560,7 +580,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 		func(kind string, diffNS bool) {
 			testEgress(kind, diffNS)
 		},
-		table.Entry("VM In Same Namespace",
+		table.Entry(fmt.Sprintf("%s: VM In Same Namespace", focusAgent),
 			reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), false),
 		table.Entry("VM In Different Namespaces",
 			reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), true),
@@ -570,7 +590,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 		func(kind string, diffNS bool) {
 			testIngress(kind, diffNS)
 		},
-		table.Entry(fmt.Sprintf("%s: VM In Same Namespaces", focusAzure),
+		table.Entry(fmt.Sprintf("%s %s: VM In Same Namespace", focusAzure, focusAgent),
 			reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), false),
 		table.Entry("VM In Different Namespaces",
 			reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), true),
@@ -585,7 +605,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			func(kind string, diffNS bool) {
 				testAppliedTo(kind, diffNS)
 			},
-			table.Entry(fmt.Sprintf("%s: VM In Same Namespace", focusAzure),
+			table.Entry(fmt.Sprintf("%s %s: VM In Same Namespace", focusAzure, focusAgent),
 				reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), false),
 			table.Entry("VM In Different Namespaces",
 				reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), true),
@@ -595,7 +615,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			func(kind string, diffNS bool) {
 				testEgress(kind, diffNS)
 			},
-			table.Entry("VM In Same Namespace",
+			table.Entry(fmt.Sprintf("%s: VM In Same Namespace", focusAgent),
 				reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), false),
 			table.Entry("VM In Different Namespaces",
 				reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), true),
@@ -605,7 +625,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 			func(kind string, diffNS bool) {
 				testIngress(kind, diffNS)
 			},
-			table.Entry(fmt.Sprintf("%s: VM In Same Namespaces", focusAzure),
+			table.Entry(fmt.Sprintf("%s %s: VM In Same Namespace", focusAzure, focusAgent),
 				reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), false),
 			table.Entry("VM In Different Namespaces",
 				reflect.TypeOf(v1alpha1.VirtualMachine{}).Name(), true),
@@ -667,7 +687,7 @@ var _ = Describe(fmt.Sprintf("%s,%s: NetworkPolicy On Cloud Resources", focusAws
 		Expect(err).NotTo(HaveOccurred())
 		err = utils.WaitApiServer(k8sClient, time.Second*60)
 		Expect(err).NotTo(HaveOccurred())
-		err = utils.CheckCloudResourceNetworkPolicies(k8sClient, kind, namespace.Name, ids, []string{anpSetupParams.Name})
+		err = utils.CheckCloudResourceNetworkPolicies(kubeCtl, k8sClient, kind, namespace.Name, ids, []string{anpSetupParams.Name}, withAgent)
 		Expect(err).ToNot(HaveOccurred())
 		err = utils.ExecuteCurlCmds(cloudVPC, nil, srcVMs, "", []string{ids[appliedIdx]}, apachePort, oks, 30)
 		Expect(err).ToNot(HaveOccurred())
